refactor: CheckInsApiService에 사용되는 Response Dto 패키지 수정
